Revert "Refactor hitbox-size from Double to Int"
This reverts commit f714fe1762988ef3262127d79224cf0c35855282.

+++ blockball-api/src/main/java/com/github/shynixn/blockball/api/persistence/entity/BallMeta.kt
@@ -53,7 +53,7 @@ interface BallMeta {
     /**
      * Size of the hitbox used for interaction detecting.
      */
-    var hitBoxSize: Int
+    var hitBoxSize: Double
 
     /**
      * Should the ball be able to carry.
